requirements.txt
beautifulsoup4==4.9.3
cachetools==4.2.1
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
dateparser==1.0.0
Flask==1.1.2
Flask-SQLAlchemy==2.5.1
google-auth==1.27.0
google-auth-oauthlib==0.4.2
GoogleNews==1.5.7
greenlet==1.0.0
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.3
joblib==1.0.1
MarkupSafe==1.1.1
numpy==1.20.2
oauthlib==3.1.0
pandas==1.2.3
psycopg2==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
python-dateutil==2.8.1
pytz==2021.1
regex==2021.3.17
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.1
scikit-learn==0.24.1
scipy==1.6.2
six==1.15.0
sklearn==0.0
soupsieve==2.2.1
SQLAlchemy==1.4.4
threadpoolctl==2.1.0
tqdm==4.59.0
tzlocal==2.1
urllib3==1.26.3
Werkzeug==1.0.1
